+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 * Fix #4256: crd-generator-apt pom.xml includes transitive dependencies
 * Fix #4294: crd-generator respects JsonIgnore annotations on enum properties
 * Fix #4247: NO_PROXY with invalid entries throws exception
+* Fix #4320: corrected leader transitions field on leader election leases
 
 #### Improvements
 * Fix #4254: adding debug logging for exec stream messages
 * Fix #4041: adding Quantity.getNumericalAmount with an explanation about bytes and cores.
 * Fix #4241: added more context to informer logs with the endpoint path
 * Fix #4250: allowing for deserialization of polymorphic unwrapped fields
+* Fix #4318: implemented LeaderElection releaseOnCancel
 * Fix #4259: Java Generator's CR should have Lombok's `@EqualsAndHashCode` with `callSuper = true`
 
 #### Dependency Upgrade